# Conflicts:
#	LineFighter/.vs/LineFighter/v15/Server/sqlite3/storage.ide-shm
#	LineFighter/.vs/LineFighter/v15/Server/sqlite3/storage.ide-wal
#	LineFighter/Library/CurrentLayout.dwlt
#	LineFighter/Library/ScriptAssemblies/Assembly-CSharp.dll.mdb
#	LineFighter/Library/StateCache/Hierarchy/579bea-mainStage.json
#	LineFighter/Library/assetDatabase3
#	LineFighter/Library/metadata/0d/0d3f8bc038127ad4395c066ed819a282
#	LineFighter/Library/metadata/0d/0d6ab16672787ce4487ce52db71212c1.info
#	LineFighter/Library/metadata/10/10145f279f90346bbb61712ff585828b
#	LineFighter/Library/metadata/14/14cc1c932e2cffe419aeb4da8eb0516b
#	LineFighter/Library/metadata/14/14cc1c932e2cffe419aeb4da8eb0516b.info
#	LineFighter/Library/metadata/16/168575773f3079f4f8b6a0396593745b
#	LineFighter/Library/metadata/16/168575773f3079f4f8b6a0396593745b.info
#	LineFighter/Library/metadata/1c/1cfd94d18b73f9747bd119c2304d3e95
#	LineFighter/Library/metadata/25/25ebe415cc14c4f4ea15c34a796b75bb
#	LineFighter/Library/metadata/3a/3aba9fe1ad29b42ffb0eea781318a626
#	LineFighter/Library/metadata/46/46c02d8d6499b4309ac1ec5d0b818585
#	LineFighter/Library/metadata/4a/4a1923e575fd3b24989e38944ecba571
#	LineFighter/Library/metadata/4a/4a1923e575fd3b24989e38944ecba571.info
#	LineFighter/Library/metadata/4b/4bcee498f4d61074aaa31f38c72ac3dc
#	LineFighter/Library/metadata/4b/4bcee498f4d61074aaa31f38c72ac3dc.info
#	LineFighter/Library/metadata/4f/4f768b64e6c37495699fffe3007e4379
#	LineFighter/Library/metadata/64/64312d09ee8d3df4ca7ea327a905f562.info
#	LineFighter/Library/metadata/93/93c1370867fcb45079592da39e0994c5
#	LineFighter/Library/metadata/97/97dff1b5ae41ee84886ae7a80ae21499
#	LineFighter/Library/metadata/a8/a8d9e44eff3f447db9448925617ccb44
#	LineFighter/Library/metadata/a9/a968e242ecd52c14a941d651cdd833dd
#	LineFighter/Library/metadata/a9/a968e242ecd52c14a941d651cdd833dd.info
#	LineFighter/Library/metadata/b2/b24d2aed911696c44b647d010cc665e5
#	LineFighter/Library/metadata/b2/b2bc8795569aa5842bc4a676301afb68
#	LineFighter/Library/metadata/bb/bbec1324b05a54939ad2fea2d7e11063
#	LineFighter/Library/metadata/c5/c5ba7626a737840ff88a3bd84b100482.info
#	LineFighter/Library/metadata/cd/cd44c1bc0fa624a39aa784af44c62180
#	LineFighter/Library/metadata/d9/d9f3a356d5ec24d41bda61dc272057a6
#	LineFighter/Library/shadercompiler-UnityShaderCompiler.exe0.log

+++ LineFighter/Assets/Scripts/Properties/Line.cs
@@ -1,10 +1,12 @@
 ﻿
+using UnityEngine;
+
 public class Line : ILine
 {
     #region Properties
     public bool LineGravity { get; set; }
 
-    public float Thickness { get; set; }
+    public double Thickness { get; set; }
 
     public int RefillRate { get; set; }
